fix logintest: moved methods inside class, removed local driver, cleaned test flow
