+++ src/test/java/POM/Homework17.java
@@ -1,3 +1,5 @@
+package POM;
+
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
