+++ src/test/java/pagefactory/Homepage.java
@@ -1,6 +1,5 @@
-package POM;
+package pagefactory;
 
-import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
