+++ src/test/resources/TestNG.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
-<suite name="TestNG"> parallel="methods" thread-count="2">
-    <parameter name="BaseURL" value="https://qa.koel.app/">
+<suite name="TestNG" parallel="methods" thread-count="2">
+    <parameter name="BaseURL" value="https://qa.koel.app/"/>
     <test name="Homework25 Test" preserve-order="false">
         <classes>
-            <class name="BaseTest"/>
+            <class name="POM.BaseTest"/>
         </classes>
     </test>
 </suite>
